[luci] Extract QuantizeBias from QuantizeWithMinMax.cpp
This extracts QuantizeBias from QuantizeWithMinMax.cpp.

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
